bot/requirements.txt
pytelegrambotapi
python-dotenv
aiohttp
firebase-admin
google-cloud-firestore

